Move ghidra helpers to new module
